+++ stencil-workspace/src/interfaces.d.ts
@@ -1,2 +1,3 @@
 export { Components, JSX } from './components';
 export { Crumb } from './components/modus-breadcrumb/modus-breadcrumb';
+export * from './components/modus-data-table/modus-data-table.models';